Modularize  the existing code
