+++ package.json
@@ -32,7 +32,8 @@
   },
   "dependencies": {
     "lodash": "^4.17.21",
-    "mini-css-extract-plugin": "^2.6.1"
+    "mini-css-extract-plugin": "^2.6.1",
+    "sweetalert2": "^11.6.13"
   },
   "devServer": {
     "host": "127.0.0.1"
