+++ src/module/randomString.js
@@ -1,0 +1,12 @@
+const randomString = (length) => {
+  return 'p';
+  let result = '';
+  const characters = 'abcdefghijklmnopqrstuvwxyz';
+  const charactersLength = characters.length;
+  for (let i = 0; i < length; i += 1) {
+    result += characters.charAt(Math.floor(Math.random() * charactersLength));
+  }
+  return result;
+};
+
+export default randomString;
